+++ ArduCopter/control_drift.cpp
@@ -46,7 +46,7 @@ void Copter::drift_run()
     static float roll_input = 0.0f;
     float target_roll, target_pitch;
     float target_yaw_rate;
-    int16_t pilot_throttle_scaled;
+    float pilot_throttle_scaled;
 
     // if not armed or motor interlock not enabled or landed and throttle at zero, set throttle to zero and exit immediately
     if(!motors.armed() || !motors.get_interlock() || (ap.land_complete && ap.throttle_zero)) {
